Compress programs and instructions by storing them in int16 and more CISC

+++ apps/hannk/interpreter/ops.h
@@ -202,13 +202,13 @@ public:
 
 class ElementwiseProgramOp : public ElementwiseOp {
 private:
-    Halide::Runtime::Buffer<int> program_;
+    Halide::Runtime::Buffer<int16_t> program_;
 
 public:
-    ElementwiseProgramOp(std::vector<TensorPtr> inputs, TensorPtr output, HalideBuffer<int> program)
+    ElementwiseProgramOp(std::vector<TensorPtr> inputs, TensorPtr output, HalideBuffer<int16_t> program)
         : ElementwiseOp(std::move(inputs), {output}), program_(program) {
     }
-    ElementwiseProgramOp(std::vector<TensorPtr> inputs, std::vector<TensorPtr> outputs, HalideBuffer<int> program)
+    ElementwiseProgramOp(std::vector<TensorPtr> inputs, std::vector<TensorPtr> outputs, HalideBuffer<int16_t> program)
         : ElementwiseOp(std::move(inputs), std::move(outputs)), program_(program) {
     }
